mta_to_dch/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o mta_to_dch/requirements.txt
certifi==2025.11.12
    # via requests
charset-normalizer==3.4.4
    # via requests
idna==3.11
    # via requests
psycopg2-binary==2.9.11
    # via mta-to-dch (pyproject.toml)
python-dotenv==1.2.1
    # via mta-to-dch (pyproject.toml)
requests==2.32.5
    # via mta-to-dch (pyproject.toml)
urllib3==2.5.0
    # via requests
